Fixed sidebar role based access
