+++ CoVAR-app/src/sidebar/sidebar.tsx
@@ -1,46 +1,31 @@
-import React, { useState, useEffect, useContext } from 'react';
-import { Link, useLocation, useNavigate } from 'react-router-dom';
-import { Box, List, ListItem, ListItemText, ListItemIcon, Typography, Button, useTheme } from '@mui/material';
-import DashboardIcon from '@mui/icons-material/Dashboard';
-import AssessmentIcon from '@mui/icons-material/Assessment';
 import AccountCircleIcon from '@mui/icons-material/AccountCircle';
-import SettingsIcon from '@mui/icons-material/Settings';
 import AdminPanelSettingsIcon from '@mui/icons-material/AdminPanelSettings';
+import AssessmentIcon from '@mui/icons-material/Assessment';
+import DashboardIcon from '@mui/icons-material/Dashboard';
 import ExitToAppIcon from '@mui/icons-material/ExitToApp';
+import GroupsIcon from '@mui/icons-material/Groups';
 import LockIcon from '@mui/icons-material/Lock';
-import GroupsIcon from '@mui/icons-material/Groups'; 
-import { getUserRole } from '../requests/requests';
+import SettingsIcon from '@mui/icons-material/Settings';
+import { Box, Button, List, ListItem, ListItemIcon, ListItemText, Typography, useTheme } from '@mui/material';
+import React, { useContext } from 'react';
+import { Link, useLocation, useNavigate } from 'react-router-dom';
 
-import { sidebarStyles, sidebarItemStyles, iconStyles, logoStyles, logoutButtonStyles } from '../styles/sidebarStyle';
-import { doSignOut } from '../firebase/auth';
 import { ThemeContext } from '../styles/customThemeProvider';
+import { iconStyles, logoStyles, logoutButtonStyles, sidebarItemStyles, sidebarStyles } from '../styles/sidebarStyle';
 
-const Sidebar: React.FC = () => {
-  const [role, setRole] = useState<string | null>(null);
+interface SidebarProps {
+  role: string | null;
+  onSignOut: () => void;
+}
+
+const Sidebar: React.FC<SidebarProps> = ({ role, onSignOut }) => {
   const navigate = useNavigate();
   const location = useLocation();
   const theme = useTheme();
 
-  useEffect(() => {
-    fetchUserRole();
-  }, []);
-
-  const fetchUserRole = async () => {
-    try {
-      const accessToken = localStorage.getItem('accessToken');
-      if (accessToken) {
-        const data = await getUserRole(accessToken);
-        setRole(data.role);
-        console.log("Role:", data.role);
-      }
-    } catch (error) {
-      console.error("Error fetching user role:", error);
-    }
-  };
-
   const signOut = async () => {
     try {
-      await doSignOut();
+      await onSignOut();
       navigate('/login');
     } catch (error) {
       console.error(error);
